+++ CHANGELOG.md
@@ -17,6 +17,7 @@ All notable changes to eww will be listed here, starting at changes since versio
 - Add `>=` and `<=` operators to simplexpr (By: viandoxdev)
 - Add `desktop` window type (By: Alvaro Lopez)
 - Add `scroll` widget (By: viandoxdev)
+- Add `notification` window type
 
 ### Notable Internal changes
 - Rework state management completely, now making local state and dynamic widget hierarchy changes possible.